SETTINGS: changed directory structre
