--- backend/apis/api_v1/api.py
@@ -1,3 +1,0 @@
-from fastapi import APIRouter
-
-api_router = APIRouter()
