+++ backend/pyproject.toml
@@ -1,0 +1,25 @@
+[tool.poetry]
+name = "sns"
+version = "0.1.0"
+description = ""
+authors = ["jeha00 <rudtls0611@naver.com>"]
+readme = "README.md"
+
+[tool.poetry.dependencies]
+python = "^3.10"
+fastapi = "^0.92.0"
+sqlalchemy = "^2.0.4"
+pymysql = "^1.0.2"
+uvicorn = "^0.20.0"
+
+
+[tool.poetry.group.test.dependencies]
+pytest = "^7.2.1"
+
+
+[tool.poetry.group.common.dependencies]
+black = "^23.1.0"
+
+[build-system]
+requires = ["poetry-core"]
+build-backend = "poetry.core.masonry.api"
