+++ backend/sns/main.py
@@ -1,0 +1,20 @@
+from fastapi import FastAPI
+from fastapi import APIRouter
+from starlette.middleware.cors import CORSMiddleware
+
+from sns.common.config import settings
+
+app = FastAPI(title=settings.PROJECT_NAME, openapi_url=f"{settings.API_V1_STR}")
+
+if settings.BACKEND_CORS_ORIGINS:
+    app.add_middleware(
+        CORSMiddleware,
+        allow_origin=[str(origin) for origin in settings.BACKEND_CORS_ORIGINS],
+        allow_credentials=True,
+        allow_methods=["*"],
+        allow_headers=["*"],
+    )
+
+
+api_router = APIRouter()
+app.include_router(api_router, prefix=settings.API_V1_STR)
