+++ backend/sns/posts/controller.py
@@ -1,0 +1,1 @@
+
