+++ backend/sns/posts/model.py
@@ -1,0 +1,1 @@
+
