+++ backend/sns/posts/schema.py
@@ -1,0 +1,1 @@
+
