+++ backend/sns/tests/model.py
@@ -1,0 +1,1 @@
+
